fix driver error in send_and_receive refactor flower structure
