+++ examples/basic/client_app.py
@@ -3,7 +3,7 @@
 from flwr.client import ClientApp, NumPyClient
 from flwr.common import Context
 
-from .task import get_dummy_model
+from task import get_dummy_model
 
 
 class FlowerClient(NumPyClient):
