--- examples/basic/flower/README.md
@@ -1,5 +1,0 @@
-Run the Example by the command
-
-```sh
-flwr run .
-```
