--- examples/basic/flower/pyproject.toml
@@ -1,32 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "quickstart"
-version = "1.0.0"
-description = ""
-license = "Apache-2.0"
-dependencies = [
-    "flwr[simulation]>=1.15.2",
-    "numpy>=2.0.2",
-]
-
-[tool.hatch.build.targets.wheel]
-packages = ["."]
-
-[tool.flwr.app]
-publisher = "abc"
-
-[tool.flwr.app.components]
-serverapp = "quickstart.server_app:app"
-clientapp = "quickstart.client_app:app"
-
-[tool.flwr.app.config]
-num-server-rounds = 3
-
-[tool.flwr.federations]
-default = "local-simulation"
-
-[tool.flwr.federations.local-simulation]
-options.num-supernodes = 10
